requirements.txt
Flask>=3.0.0
gunicorn>=21.2.0
python-dotenv>=1.0.1
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
nba_api>=1.5.0
requests>=2.31.0